make AllConversation.kt use of composite function
