+++ app/src/main/kotlin/com/wire/android/ui/main/conversationlist/AllConversation.kt
@@ -1,30 +1,22 @@
 package com.wire.android.ui.main.conversationlist
 
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.unit.dp
 import com.wire.android.R
 import com.wire.android.ui.common.UserProfileAvatar
-import com.wire.android.ui.main.conversationlist.common.EventBadgeFactory
+import com.wire.android.ui.main.conversationlist.common.ConversationItem
 import com.wire.android.ui.main.conversationlist.common.GroupConversationAvatar
-import com.wire.android.ui.main.conversationlist.common.RowItem
+import com.wire.android.ui.main.conversationlist.common.GroupName
 import com.wire.android.ui.main.conversationlist.common.UserLabel
 import com.wire.android.ui.main.conversationlist.common.folderWithElements
 import com.wire.android.ui.main.conversationlist.model.Conversation
 import com.wire.android.ui.main.conversationlist.model.Conversation.GroupConversation
 import com.wire.android.ui.main.conversationlist.model.Conversation.PrivateConversation
 import com.wire.android.ui.main.conversationlist.model.ConversationFolder
+import com.wire.android.ui.main.conversationlist.model.EventType
 import com.wire.android.ui.main.conversationlist.model.NewActivity
 import com.wire.android.ui.main.conversationlist.model.toUserInfoLabel
-import com.wire.android.ui.theme.body02
 
 
 @Composable
@@ -52,10 +44,12 @@ private fun AllConversationContent(
             header = { stringResource(id = R.string.conversation_label_new_activity) },
             items = newActivities
         ) { newActivity ->
-            NewActivityRowItem(
-                newActivity = newActivity,
-                onConversationItemClick
-            )
+            with(newActivity) {
+                AllConversationItem(
+                    conversation = conversation,
+                    eventType = eventType
+                )
+            }
         }
 
         conversations.forEach { (conversationFolder, conversationList) ->
@@ -63,9 +57,8 @@ private fun AllConversationContent(
                 header = { conversationFolder.folderName },
                 items = conversationList
             ) { conversation ->
-                ConversationItem(
+                AllConversationItem(
                     conversation = conversation,
-                    onConversationItemClick
                 )
             }
         }
@@ -73,54 +66,29 @@ private fun AllConversationContent(
 }
 
 @Composable
-private fun NewActivityRowItem(
-    newActivity: NewActivity,
-    onConversationItemClick: () -> Unit
-) {
-    RowItem(onRowItemClick = onConversationItemClick) {
-        createConversationLabel(conversation = newActivity.conversation)
-        Box(modifier = Modifier.fillMaxWidth()) {
-            EventBadgeFactory(
-                eventType = newActivity.eventType,
-                modifier = Modifier
-                    .align(Alignment.CenterEnd)
-                    .padding(end = 8.dp)
-            )
-        }
-    }
-}
-
-@Composable
-private fun createConversationLabel(conversation: Conversation) {
+private fun AllConversationItem(conversation: Conversation, eventType: EventType? = null) {
     when (conversation) {
         is GroupConversation -> {
             with(conversation) {
-                GroupConversationAvatar(groupColorValue)
-                GroupName(groupName)
+                ConversationItem(
+                    avatar = {
+                        GroupConversationAvatar(colorValue = groupColorValue)
+                    },
+                    title = { GroupName(groupName) },
+                    eventType = eventType
+                )
             }
         }
         is PrivateConversation -> {
-            with(conversation) {
-                UserProfileAvatar(avatarUrl = userInfo.avatarUrl, onClick = {})
-                UserLabel(toUserInfoLabel())
-            }
+            ConversationItem(
+                avatar = {
+                    UserProfileAvatar()
+                },
+                title = { UserLabel(conversation.toUserInfoLabel()) },
+                eventType = eventType
+            )
         }
     }
-}
-
-@Composable
-private fun ConversationItem(conversation: Conversation, onConversationItemClick: () -> Unit) {
-    RowItem(onRowItemClick = onConversationItemClick) {
-        createConversationLabel(conversation = conversation)
-    }
-}
-
-@Composable
- fun GroupName(name: String) {
-    Text(
-        text = name,
-        style = MaterialTheme.typography.body02
-    )
 }
